standardrb
I can't take Rubocop any more. Moving up to Ruby 3 forces us to update RuboCop,
and it comes with dozens of linters that I'd have to evaluate, none of which is
a marginal improvement. I'm done having opinions.

+++ app/models/hidden_story.rb
@@ -5,14 +5,12 @@ class HiddenStory < ApplicationRecord
   scope :by, ->(user) { where(user: user) }
 
   def self.hide_story_for_user(story_id, user_id)
-    HiddenStory.where(:user_id => user_id, :story_id =>
-      story_id).first_or_initialize.save!
+    HiddenStory.where(user_id: user_id, story_id: story_id).first_or_initialize.save!
     ReadRibbon.hide_replies_for(story_id, user_id)
   end
 
   def self.unhide_story_for_user(story_id, user_id)
-    HiddenStory.where(:user_id => user_id, :story_id =>
-      story_id).delete_all
+    HiddenStory.where(user_id: user_id, story_id: story_id).delete_all
     ReadRibbon.unhide_replies_for(story_id, user_id)
   end
 end
